refactor(rlp): introduce `Fields` type
